pyproject.toml
[tool.poetry]
name = "pdf-summarizer-end-to-end-project"
version = "0.1.0"
description = ""
authors = ["chm10 <christian.maekawa@gmail.com>"]
readme = "README.md"
packages = [{include = "pdf_summarizer_end_to_end_project"}]

[tool.poetry.dependencies]
python = "^3.11"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"
streamlit = "^1.33.0"
streamlit-extras = "^0.4.2"
faiss-cpu = "^1.8.0"
tiktoken = "^0.6.0"
langchain-community = "^0.0.34"
langchain-openai = "^0.1.3"
langchain = "^0.1.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
